sentinel_rag/requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.1.0
langchain-groq>=0.1.0
langchain-openai>=0.0.8
langchain-huggingface>=0.0.3
sentence-transformers>=3.0.0
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
openai>=1.12.0
pypdf>=4.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
tiktoken>=0.6.0
pydantic>=2.0.0
numpy>=1.26.0
